Fixed bug --> Remove with no checkboxes selected: added @RequestParam(required = false)
Also fixed error thrown in browser --> jQuery null

+++ src/main/java/org/launchcode/controllers/CheeseController.java
@@ -66,9 +66,11 @@ public class CheeseController {
     }
 
     @RequestMapping(value = "remove", method = RequestMethod.POST)
-    public String processRemoveCheeseForm(@RequestParam int [] cheeseIds) {
-        for (int cheeseId : cheeseIds) {
-            cheeseDao.delete(cheeseId);
+    public String processRemoveCheeseForm(@RequestParam(required = false) int [] cheeseIds) {
+        if (cheeseIds != null) {
+            for (int cheeseId : cheeseIds) {
+                cheeseDao.delete(cheeseId);
+            }
         }
         return "redirect:";
     }
